Make sure warnings aren't triggered by no IDs.
Make sure that other plugins that try to filter `the_title` without passing in two parameters don't interfere with the Subtitles plugin. When `the_title` is filtered, two parameters are needed by Subtitles: the title of the post and the post ID.

If a custom widget or anything else that uses `the_title` does not provide this, then Subtitles will return a PHP warning. So, set the default ID to null if it hasn't been passed to `the_subtitle` and all should be well.